Reorganize PlannerNote request
