+++ Core/CoreTests/Planner/Model/API/APIPlannableTests.swift
@@ -112,33 +112,6 @@ class APIPlannableTests: XCTestCase {
     }
 }
 
-class PostPlannerNoteRequestTests: XCTestCase {
-    var req: PostPlannerNoteRequest!
-    var date: Date = Clock.now
-
-    override func setUp() {
-        super.setUp()
-        req = PostPlannerNoteRequest(body: PostPlannerNoteRequest.Body(title: "title", details: "details", todo_date: date, course_id: "1", linked_object_type: .planner_note, linked_object_id: "1"))
-    }
-
-    func testMethod() {
-        XCTAssertEqual(req.method, .post)
-    }
-
-    func testPath() {
-        XCTAssertEqual(req.path, "planner_notes")
-    }
-
-    func testBody() {
-        XCTAssertEqual(req.body?.title, "title")
-        XCTAssertEqual(req.body?.details, "details")
-        XCTAssertEqual(req.body?.todo_date, date)
-        XCTAssertEqual(req.body?.course_id, "1")
-        XCTAssertEqual(req.body?.linked_object_type, .planner_note)
-        XCTAssertEqual(req.body?.linked_object_id, "1")
-    }
-}
-
 class UpdatePlannerOverrideRequestTests: XCTestCase {
     private let testee = UpdatePlannerOverrideRequest(overrideId: "testID", body: .init(marked_complete: true))
